separate sign up and login handlers to different files
